[WIP] Migrating all Post components over to typescript
