+++ src/components/Post/Author/Author.tsx
@@ -1,4 +1,3 @@
-// @flow strict
 import React from 'react';
 import { getContactHref } from '../../../utils';
 import styles from './Author.module.scss';
