+++ src/components/Post/Comments/Comments.tsx
@@ -1,11 +1,10 @@
-// @flow strict
 import React from 'react';
 import ReactDisqusComments from 'react-disqus-comments';
 import { useSiteMetadata } from '../../../hooks';
 
 type Props = {
-  postTitle: string,
-  postSlug: string
+  postTitle: string;
+  postSlug: string;
 };
 
 const Comments = ({ postTitle, postSlug }: Props) => {
@@ -15,14 +14,7 @@ const Comments = ({ postTitle, postSlug }: Props) => {
     return null;
   }
 
-  return (
-    <ReactDisqusComments
-      shortname={disqusShortname}
-      identifier={postTitle}
-      title={postTitle}
-      url={url + postSlug}
-    />
-  );
+  return <ReactDisqusComments shortname={disqusShortname} identifier={postTitle} title={postTitle} url={url + postSlug} />;
 };
 
 export default Comments;
