+++ src/components/Post/Content/index.ts
@@ -1,2 +1,1 @@
-// @flow strict
 export { default } from './Content';
