+++ src/components/Post/Meta/Meta.tsx
@@ -1,10 +1,9 @@
-// @flow strict
 import React from 'react';
 import moment from 'moment';
 import styles from './Meta.module.scss';
 
 type Props = {
-  date: string
+  date: string;
 };
 
 const Meta = ({ date }: Props) => (
